requirements.txt
lxml==4.6.3
pdf2image==1.15.1
Pillow==8.2.0
python-pptx==0.6.19
XlsxWriter==1.4.3
